+++ package/naruse-h5/components/image/index.tsx
@@ -2,10 +2,11 @@ import { commonEventHander } from '../../core/event';
 import React, { Component } from 'react';
 import cssStyle from './index.css'
 import { getPropsDataSet } from '../../utils';
+import { getNaruseConfig } from '../../core/init';
 
 const h = React.createElement;
 
-class Image extends Component<
+class _Image extends Component<
 {
     lazyLoad,
     mode,
@@ -27,10 +28,12 @@ class Image extends Component<
     ref: HTMLImageElement | null;
     constructor(props) {
         super(props);
-        this.state = { isLoaded: false };
+        const { unsafeEnabled: { compatibleWeexElement } } = getNaruseConfig();
+        this.state = { isLoaded: false, imageSize: {}, visible: !compatibleWeexElement };
         this.imageOnLoad = this.imageOnLoad.bind(this);
         this.observer = {};
         this.ref = null;
+        this.naturalSize = null;
     }
     componentDidMount () {
         if (this.props.lazyLoad) {
@@ -44,14 +47,79 @@ class Image extends Component<
             }, { rootMargin: '300px 0px' });
             this.observer.observe(this.ref);
         }
+        this.computeNaturalSize(() => {
+            this.adaptationWeex();
+            this.setState({ visible: true });
+        });
     }
 
+    computeNaturalSize (callback) {
+        const image = new Image();
+        image.src = this.props.src;
+        image.onload = () => {
+            this.naturalSize = { width: image.width, height: image.height };
+            callback(true);
+        };
+        image.onerror = () => {
+            callback(false);
+            this.naturalSize = null;
+        };
+        image.onabort = () => {
+            callback(false);
+            this.naturalSize = null;
+        };
+    }
+    componentDidUpdate(prevProps) {
+        if(prevProps.src != this.props.src) {
+            this.computeNaturalSize(() => {
+                this.adaptationWeex();
+            });
+        } else if (
+            prevProps.style?.width !== this.props.style?.width
+            || prevProps.style?.height !== this.props.style?.height
+        ) {
+            this.adaptationWeex();
+        }
+    }
+    adaptationWeex () {
+        const { style = {}, mode } = this.props;
+        const { unsafeEnabled: { compatibleWeexElement, compatibleWeexElementLog }, convertRpx } = getNaruseConfig();
+        if (compatibleWeexElementLog) {
+            console.log('compatibleWeexElement:', this.props, this.naturalSize);
+        }
+        if (
+            compatibleWeexElement
+            && (style.height == undefined && style.width === undefined)
+            && this.naturalSize
+        ) {
+            switch (mode) {
+                case 'widthFix':
+                    this.setState({
+                        imageSize: {
+                            maxWidth: convertRpx(this.naturalSize.width) + 'px',
+                        }
+                    });
+                    return;
+                case 'heightFix':
+                    this.setState({
+                        imageSize: {
+                            maxHeight: convertRpx(this.naturalSize.height) + 'px',
+                        }
+                    });
+                    return;
+            }
+        } else {
+            this.setState({ imageSize: {} });
+        }
+    }
     componentWillUnmount () {
         this.observer.disconnect && this.observer.disconnect();
     }
 
     /** 当图片加载完毕 */
-    imageOnLoad = commonEventHander.bind(this)
+    imageOnLoad = (event) => {
+        commonEventHander.call(this, event);
+    }
 
     render () {
         const {
@@ -64,26 +132,29 @@ class Image extends Component<
             id,
             ...other
         } = this.props;
+        const { imageSize, visible } = this.state;
+        if (!visible) {
+            return null;
+        }
         const divStyle = { ...cssStyle.naruseImg, ...(mode === 'widthFix' ? cssStyle.naruseImg__widthfix : {}) };
         const imgStyle = cssStyle[(mode || 'scaleToFill').toLowerCase().replace(/\s/g, '')];
         return (
             <div onClick={commonEventHander.bind(this)} className={className} style={{ ...divStyle, ...style }}>
-                {
-                    <img
-                        ref={img => (this.ref = img)}
-                        id={id}
-                        style={imgStyle}
-                        src={src}
-                        onLoad={this.imageOnLoad}
-                        onError={onError}
-                        onTransitionEnd={commonEventHander.bind(this)}
-                        {...imgProps}
-                        {...getPropsDataSet(other)}
-                    />
-                }
+                <img
+                    key='img'
+                    ref={img => (this.ref = img)}
+                    id={id}
+                    style={{ ...imageSize, ...imgStyle }}
+                    src={src}
+                    onLoad={this.imageOnLoad}
+                    onError={onError}
+                    onTransitionEnd={commonEventHander.bind(this)}
+                    {...imgProps}
+                    {...getPropsDataSet(other)}
+                />
             </div>
         );
     }
 }
 
-export default Image;
+export default _Image;