+++ package/naruse-h5/core/init.ts
@@ -1,8 +1,8 @@
-import { NaruseConfig, NaruseInitParams } from "../../naruse-share";
+import { NaruseConfigH5, NaruseInitParamsH5 } from '../../naruse-share';
 import { logger } from "../utils/log";
 import { withPageInit } from "./withPage";
 
-const _config: NaruseConfig = {
+const _config: NaruseConfigH5 = {
     hotPuller: () => {
         logger.error('未初始化热更新拉取，无法更新组件默认为空');
         return Promise.resolve({ code: '', ctx: {} });
@@ -18,6 +18,10 @@ const _config: NaruseConfig = {
     hotImport: (_path, _ctx) => {
         logger.error('hotImport 函数尚未初始化！');
         return Promise.resolve('');
+    },
+    unsafeEnabled: {
+        compatibleWeexElement: false,
+        compatibleWeexElementLog: false,
     }
 };
 
@@ -35,14 +39,13 @@ const getNaruseConfig = () => {
  * @description naruse内部初始化过程
  * @author CHC
  * @date 2022-06-14 10:06:36
- * @param hotPuller 热更新处理、广告加载
- * @param baseCtx 广告运行时的上下文环境
- * @param onRunError 广告运行错误时触发
- * @param convertRpx 自定义 rpx 到 px 的转换
- * @param pageContainer 能获取到页面滚动条偏移量的容器元素
+ * @param newConfig
  */
-const naruseInit = (newConfig: NaruseInitParams) => {
+const naruseInit = (newConfig: NaruseInitParamsH5) => {
+    const unsafeEnabled = newConfig.unsafeEnabled;
+    delete newConfig.unsafeEnabled;
     Object.assign(_config, newConfig);
+    Object.assign(_config.unsafeEnabled, unsafeEnabled);
     const { pageContainer } = _config as any;
     withPageInit({ pageContainer });
 }